requirements.txt
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.4
cycler==0.10.0
decorator==4.4.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
monty==2.0.6
mpmath==1.1.0
networkx==2.3
nose2==0.9.1
numpy==1.17.2
palettable==3.3.0
pandas==0.25.1
Pillow==6.1.0
PyDispatcher==2.0.5
pymatgen==2019.9.16
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
scipy==1.3.1
six==1.12.0
spglib==1.14.1.post0
sympy==1.4
tabulate==0.8.3
urllib3==1.25.3
